+++ crates/cuda_std/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Notable changes to this project will be documented in this file.
 
+## Unreleased
+
+- Added warp shuffles, matches, reductions, and votes in the `warp` module.
+- Added `activemask` in the `warp` module to query a mask of the active threads.
+
 ## 0.2.2 - 2/7/22
 
 - Thread/Block/Grid index/dim intrinsics now hint to llvm that their range is in some bound declared by CUDA. Hopefully allowing for more optimizations.